refactor: inline manager relationships
